Removed Failed tests
