+++ tests/test_admin.py
@@ -1,8 +1,6 @@
 from unittest.mock import patch
-
 from admin_auto_tests.test_model import ModelAdminTestCase
 from model_mommy import mommy, random_gen
-
 from tools.models import Organization
