+++ tests/test_ooi_list.py
@@ -1,6 +1,4 @@
 from unittest.mock import Mock
-import pytest
-from django.contrib.auth.models import Permission, ContentType
 from django.contrib.messages.middleware import MessageMiddleware
 from django.contrib.sessions.middleware import SessionMiddleware
 from django.urls import reverse, resolve
@@ -11,31 +9,6 @@ from octopoes.models.pagination import Paginated
 from octopoes.models.types import OOIType, Network
 from pytest_django.asserts import assertContains
 from rocky.views import OOIListView
-from tools.models import OrganizationMember
-
-
-@pytest.fixture
-def my_user(user, organization):
-    OrganizationMember.objects.create(
-        user=user,
-        organization=organization,
-        verified=True,
-        authorized=True,
-        status=OrganizationMember.STATUSES.ACTIVE,
-        trusted_clearance_level=4,
-        acknowledged_clearance_level=4,
-    )
-    content_type = ContentType.objects.get_by_natural_key("tools", "organizationmember")
-    permission, _ = Permission.objects.get_or_create(
-        content_type=content_type,
-        codename="can_scan_organization",
-    )
-    user.user_permissions.add(permission)
-
-    device = user.staticdevice_set.create(name="default")
-    device.token_set.create(token=user.get_username())
-
-    return user
 
 
 def setup_octopoes_mock() -> Mock:
@@ -44,8 +17,6 @@ def setup_octopoes_mock() -> Mock:
     return mock
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
 def setup_request(request, user, organization):
     """
     Setup request with middlewares, user, organization and octopoes
@@ -63,8 +34,6 @@ def setup_request(request, user, organization):
     return request
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
 def test_ooi_list(rf, my_user, organization):
     url = reverse("ooi_list", kwargs={"organization_code": organization.code})
     request = rf.get(url)
@@ -82,8 +51,6 @@ def test_ooi_list(rf, my_user, organization):
     assertContains(response, "testnetwork")
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
 def test_ooi_list_with_clearance_type_filter_and_clearance_level_filter(rf, my_user, organization):
     request = rf.get(
         reverse("ooi_list", kwargs={"organization_code": organization.code}),
