+++ tests/test_upload_csv.py
@@ -1,36 +1,11 @@
 from io import BytesIO
 import pytest
-from django.contrib.auth.models import Permission, ContentType
 from django.contrib.messages.middleware import MessageMiddleware
 from django.contrib.sessions.middleware import SessionMiddleware
 from django.urls import reverse
 from django_otp import DEVICE_ID_SESSION_KEY
 from django_otp.middleware import OTPMiddleware
 from rocky.views import UploadCSV
-from tools.models import OrganizationMember
-
-
-@pytest.fixture
-def my_user(user, organization):
-    OrganizationMember.objects.create(
-        user=user,
-        organization=organization,
-        verified=True,
-        authorized=True,
-        status=OrganizationMember.STATUSES.ACTIVE,
-        trusted_clearance_level=4,
-        acknowledged_clearance_level=4,
-    )
-    permission, _ = Permission.objects.get_or_create(
-        content_type=ContentType.objects.get_by_natural_key("tools", "organizationmember"),
-        codename="can_scan_organization",
-    )
-    user.user_permissions.add(permission)
-
-    device = user.staticdevice_set.create(name="default")
-    device.token_set.create(token=user.get_username())
-
-    return user
 
 
 CSV_EXAMPLES = [
@@ -58,9 +33,7 @@ INPUT_TYPES = ["Hostname", "Hostname", "IPAddressV4", "IPAddressV6", "URL", "URL
 EXPECTED_OOI_COUNTS = [2, 2, 6, 4, 4, 2]
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
-def test_upload_csv_page(rf, client, my_user, organization):
+def test_upload_csv_page(rf, my_user, organization):
     request = rf.get(reverse("upload_csv", kwargs={"organization_code": organization.code}))
     request.user = my_user
     request.organization = organization
@@ -69,9 +42,7 @@ def test_upload_csv_page(rf, client, my_user, organization):
     assert response.status_code == 200
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
-def test_upload_csv_simple(rf, client, my_user, organization):
+def test_upload_csv_simple(rf, my_user, organization):
     request = rf.get(reverse("upload_csv", kwargs={"organization_code": organization.code}))
     request.user = my_user
     request.organization = organization
@@ -84,9 +55,7 @@ def test_upload_csv_simple(rf, client, my_user, organization):
     assert response.status_code == 200
 
 
-@pytest.mark.django_db
-@pytest.mark.usefixtures("organization")
-def test_upload_bad_input(rf, client, my_user, organization, mocker):
+def test_upload_bad_input(rf, my_user, organization, mocker):
     mocker.patch("rocky.views.upload_csv.save_ooi")
 
     example_file = BytesIO(b"invalid|'\n4\bcsv|format")
@@ -116,7 +85,7 @@ def test_upload_bad_input(rf, client, my_user, organization, mocker):
     "example_input, input_type, expected_ooi_counts",
     [CSV_EXAMPLES, INPUT_TYPES, EXPECTED_OOI_COUNTS],
 )
-def test_upload_csv(rf, client, my_user, mocker, organization, example_input, input_type, expected_ooi_counts):
+def test_upload_csv(rf, my_user, mocker, organization, example_input, input_type, expected_ooi_counts):
     mock_save_ooi = mocker.patch("rocky.views.upload_csv.save_ooi")
 
     example_file = BytesIO(example_input)
